[fx] Cache translation_validation_enabled on `ShapeEnv`
`ShapeEnv` has tons of functionallity that is conditioned on this
`translation_validation_enabled()` check, to the point where 8% of
time in `empty_strided` is spent just in that function.

However, it doesn't really make sense for the value of
`translation_validation_enabled()` to change throughout the life of a `ShapeEnv`
so we might as well run the check once and store it in the `ShapeEnv`.

ghstack-source-id: 5849c46e9a6b6724f229f65a516633127e912bdf
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112493