Allow binary pointwise operations to cause refinement on unbacked SymInts
To do this, there is a little detour to remove hint caching for unbacked
SymInts; now, we just always attempt to update the hint (using
maybe_evaluate_static; this is much better than the replace we were
doing before) if we don't think we know it.

With this change, we now can generally infer that i0 == 1 is false for
a size-like unbacked SymInt.  So if we write the size match /
broadcasting test very carefully (see comment), we will eventually
end up expect_true(sizeA == sizeB), which is good enough to cause
refinement.  Phew!

I think I still want to setup a replacement if you do i0 == s0, but I'm
going to do that in a follow up.

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: c52dd9ae0213dd6eb70fe0fc5644c6e9626f665e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112155