Set default for IS_FBCODE flag
Summary:
If IS_FBCODE is False, then we print an OSS repro if a test fails. We do
set IS_FBCODE manually on most internal tests, but we don't do it for
all of them. This PR changes it so that the IS_FBCODE gets set to the
correct default value (and then tests are able to override them if
they'd like).

Test Plan:
- Tested locally

ghstack-source-id: 9f9368ddfdefc03e1059cfcad41c9c0e2725dcad
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112766